- Исправления 1.16.3 - Причесал весь проект по MVVM.
