+++ Tracker/Models/CoreData/Tracker.swift
@@ -11,7 +11,6 @@ struct Tracker {
     let id: UUID
     let name: String
     let color: String
-//    let color: UIColor
     let emoji: String
     let schedule: String
 }
@@ -20,15 +19,7 @@ extension Tracker {
     init(coreDataObject: TrackerCoreData) {
         self.id = coreDataObject.id  ?? UUID()
         self.name = coreDataObject.name ?? ""
-        
         self.color = coreDataObject.colorHex ?? "#000000"
-        
-//        if let colorHex = coreDataObject.colorHex {
-//            self.color = UIColor(hex: colorHex)
-//        } else {
-//            self.color = UIColor.green
-//        }
-        
         self.emoji = coreDataObject.emoji ?? ""
         self.schedule = coreDataObject.schedule ?? ""
     }
