+++ Tracker/Models/TrackerCoreManager.swift
@@ -128,7 +128,6 @@ final class TrackerCoreManager: NSObject {
             if let foundCategory = result.first {
                 guard let color = newTracker.trackers.first?.color else { return }
                 let colorInString = color
-//                    .hexStringFromUIColor()
                 
                 let newTrackerToAdd = TrackerCoreData(context: context)
                 newTrackerToAdd.id = newTracker.trackers.first?.id
@@ -147,7 +146,6 @@ final class TrackerCoreManager: NSObject {
             
             guard let color = newTracker.trackers.first?.color else { return }
             let colorInString = color
-//                .hexStringFromUIColor()
             
             let newTrackerToAdd = TrackerCoreData(context: context)
             newTrackerToAdd.id = newTracker.trackers.first?.id
