+++ Tracker/ViewControllers/CreatingNewTrackerVC/Views/CreatingNewTrackerViewController.swift
@@ -1,5 +1,5 @@
 //
-//  CreatingOneOffTrackerVC.swift
+//  CreatingNewTrackerViewController.swift
 //  Tracker
 //
 //  Created by Kirill Sklyarov on 13.03.2024.
@@ -7,12 +7,13 @@
 
 import UIKit
 
-final class CreatingOneOffTrackerVC: UIViewController {
+final class CreatingNewTrackerViewController: UIViewController {
     
     // MARK: - UI Properties
     let trackerNameTextField = UITextField()
     let tableView = UITableView()
     let exceedLabel = UILabel()
+    
     let emojiCollection = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
     let colorsCollection = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
     
@@ -22,7 +23,7 @@ final class CreatingOneOffTrackerVC: UIViewController {
     let contentStackView = UIStackView()
     
     // MARK: - Private Properties
-    let viewModel = CreatingOneOffTrackerViewModel()
+    let viewModel = CreatingNewTrackerViewModel()
     
     // MARK: - Life Cycles
     override func viewDidLoad() {
@@ -46,17 +47,15 @@ final class CreatingOneOffTrackerVC: UIViewController {
     }
     
     @objc func createButtonTapped(_ sender: UIButton) {
-        guard let trackerName = trackerNameTextField.text else {
-            print("We have a problem here")
-            return
-        }
-        viewModel.createNewTracker(trackerNameTextField: trackerName)
+        guard let name = trackerNameTextField.text else { print("Что-то пошло не так"); return }
+        viewModel.createNewTask(trackerNameTextField: name)
     }
     
     // MARK: - Private Methods
     private func setupUI() {
         
         setupTextField()
+        setupContentStack()
         
         setupTableView()
         
@@ -64,10 +63,10 @@ final class CreatingOneOffTrackerVC: UIViewController {
         
         setupColorsCollectionView()
         
-        setupScrollView()
-        
-        self.title = "Новое нерегулярное событие"
+        self.title = "Новая привычка"
         view.backgroundColor = UIColor(named: "projectBackground")
+        
+        setupScrollView()
     }
     
     private func setupButtons(title: String) -> UIButton {
@@ -83,7 +82,46 @@ final class CreatingOneOffTrackerVC: UIViewController {
         return button
     }
     
-    // MARK: - Public Methods
+    
+    private func setupTextField() {
+        
+        let rightPaddingView = UIView()
+        let clearTextFieldButton: UIButton = {
+            let button = UIButton(type: .custom)
+            let configuration = UIImage.SymbolConfiguration(pointSize: 17)
+            let imageColor = UIColor(named: "createButtonGrayColor") ?? .lightGray
+            let image = UIImage(systemName: "xmark.circle.fill", withConfiguration: configuration)?
+                .withRenderingMode(.alwaysOriginal)
+                .withTintColor(imageColor)
+            button.setImage(image, for: .normal)
+            button.addTarget(self, action: #selector(clearTextButtonTapped), for: .touchUpInside)
+            return button
+        } ()
+        
+        lazy var clearTextStack: UIStackView = {
+            let stack = UIStackView()
+            stack.axis = .horizontal
+            stack.addArrangedSubview(clearTextFieldButton)
+            stack.addArrangedSubview(rightPaddingView)
+            stack.translatesAutoresizingMaskIntoConstraints = false
+            stack.widthAnchor.constraint(equalToConstant: 28).isActive = true
+            return stack
+        } ()
+        
+        trackerNameTextField.placeholder = "Введите название трекера"
+        let leftPaddingView = UIView(frame: CGRect(x: 0, y: 0, width: 10, height: 75))
+        trackerNameTextField.leftView = leftPaddingView
+        trackerNameTextField.leftViewMode = .always
+        trackerNameTextField.rightView = clearTextStack
+        trackerNameTextField.rightViewMode = .whileEditing
+        trackerNameTextField.textAlignment = .left
+        trackerNameTextField.layer.cornerRadius = 10
+        trackerNameTextField.backgroundColor = UIColor(named: "textFieldBackgroundColor")
+        trackerNameTextField.heightAnchor.constraint(equalToConstant: 75).isActive = true
+        
+        trackerNameTextField.delegate = self
+    }
+    
     func isCreateButtonEnable() {
         isAllFieldsFilled() ? createButtonIsActive() : createButtonIsNotActive()
     }
